messaging_app/requirements.txt
Django==5.2.8
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-filter==24.3
djangorestframework-nested==0.0.1